+++ src/pipcrop.js
@@ -1119,7 +1119,7 @@ var PipCrop = (function () {
   }
 
   const api = {
-    version: "1.2.0",
+    version: "1.3.0",
     install,
     uninstall,
     get installed() {